refactor: split shared schema
