+++ drizzle.config.ts
@@ -8,7 +8,7 @@ if (!process.env.DATABASE_URL) {
 
 export default defineConfig({
   out: "./migrations",
-  schema: "./shared/schema.ts",
+  schema: "./shared/schema",
   dialect: "postgresql",
   dbCredentials: {
     url: process.env.DATABASE_URL,
