+++ shared/schema/index.ts
@@ -1,0 +1,9 @@
+export * from './users';
+export * from './candidates';
+export * from './employers';
+export * from './jobPosts';
+export * from './applications';
+export * from './shortlists';
+export * from './matchScores';
+export * from './adminInviteCodes';
+export * from './searchAnalytics';
