+++ shared/zod/adminInviteCodes.ts
@@ -1,0 +1,4 @@
+import { createInsertSchema } from 'drizzle-zod';
+import { adminInviteCodes } from '../schema';
+
+export const insertAdminInviteCodeSchema = createInsertSchema(adminInviteCodes);
