+++ shared/zod/candidates.ts
@@ -1,0 +1,10 @@
+import { createInsertSchema } from 'drizzle-zod';
+import { candidates } from '../schema';
+
+export const insertCandidateSchema = createInsertSchema(candidates).omit({
+  id: true,
+  profileStatus: true,
+  deleted: true,
+  createdAt: true,
+  updatedAt: true,
+});
