+++ shared/zod/searchAnalytics.ts
@@ -1,0 +1,4 @@
+import { createInsertSchema } from 'drizzle-zod';
+import { searchAnalytics } from '../schema';
+
+export const insertSearchAnalyticsSchema = createInsertSchema(searchAnalytics);
